Fix #1: Replace Flatfile DB with something that is maintained
The author of the PHP Flatfile package recommends not to use it
anymore, but to use SQLite instead. And indeed, for Realblog_XH's needs
it actually makes sense to use an SQL database opposed to some NoSQL
database in the long run.

We don't feel the need to use an abstraction layer, so we choose the
SQLite3 package which is supposed to be available on many shared
hosting servers.

We do not make any special use of the advanced capabilities of an SQL
database for now, though.