Fix database creation (fixes #15)
We still had the old code which created the Flatfile DB files, which is
of course superfluous after switching to SQLite3. Also we have to make
sure that the content folder is created before trying to create the
SQlite3 database. Furthermore we must not attempt to import the old
Flatfile DB, if it doesn't exist.