Fix #2: Store categories in database
We store the categories in a separate field in the database instead of
as CMSimple scripting in the teaser or header. That is more user
friendly, and the filtering is faster. Actually, we should normalize
the categories, but we leave that for another time.

We adapt the import functionality to extract categories stored in the
teaser and/or body into the categories field.

We believe that we can now remove the limitation note in the manual;
most likely users can now blog to their heart's content. Even on an
average desktop machine 10,000 articles don't appear to have
performance issues.