Check that XH_ADM is defined
Under old CMSimple_XH versions XH_ADM is not defined, so checking for
XH_ADM would succeed due to PHP's fallback to take it as string. We had
catered to this issue by checking for a supported version in index.php
which also prevented direct access to this file. However, this is
actually unnecessary, so we get rid of this code and add explicit
checks whether XH_ADM is defined before we access it.