Add range constraint to status column
We add a range constraint on the status column of the articles table to
avoid inadvertent values. This allows us to remove the explicit check
in the status bulk update functionality. We also remove the
backup/saved status, which doesn't make much sense – presumably it has
been introduced to avoid auto-publishing drafts, but this will have to
be resolved more generally anyway (see #13).

We also make sure that the status is valid when importing old Flatfile
DBs, and remove the language string which is not needed anymore.