Add optimistic offline lock
We add a version column to the article table to ensure that articles
which have been modified in the meantime can neither be saved or
deleted. As concurrency is assumed to be pretty low, we do not give a
special error message in the case that the offline lock triggers, let
alone that we offer some conflict resolution.

The bulk operations (deletion and status change) are not subject to the
lock by their very nature.