Close #32: Remove config option teaser→multicolumns
Actually, this is supposed to be a CSS only issue, so we don't need any
configuration option to alter the behavior. Instead, we add
div.realblog_entry_preview around each article preview, and prepare the
CSS for easy adoption.

The hack to force the height of .realblog_entries_preview to contain
its floating children has been shamelessly stolen from
http://stackoverflow.com/questions/218760#11597829. Thanks!