Fix #13: Status changes wrt. auto publishing and archiving
As it was, the publishing date of new articles defaulted to 1970-01-01.
If auto-publishing was enabled, the article would have been published
immediately after saving what could cause confusion. Even worse, if
auto-publishing was disabled, but would have been enabled later on,
all formerly unpublished articles would have been automatically
published. We fix that by simply defaulting to 2038-01-19 for new
articles.

A related issue was that the publishing and archiving dates were reset,
if auto-publishing and auto-archiving had been disabled, and the
article would have been saved. We keep the manually set dates now, just
in case the user wants to switch back to auto-publishing and
auto-archiving.