Close #29: Category selection
We progressively enhance the categories text input by adding a datalist
of all existing categories. This works quite well for a single category
(depending on the browser), but doesn't really fit well for multiple
categories. Nonetheless, we consider this enhancement a worthwhile
interim stage before we're able to normalize the categories and offer
a separate category administration.