Fix #45: Time of RSS <pubDate> is always 00:00:00
While mostly dates without times are sufficient for Realblog_XH, the
`<pubDate>` elements of RSS feeds expect a date/time in RFC 822 format.
To let this have a somewhat meaningful time component, we make sure to
transparently include the current time whenever an article's date is
changed, what includes the creation of a new article.