Close #26: Option to show teaser as lead above article
We implement the new config option show_teaser which allows to show the
teaser as lead above the article. This offers the possibility to treat
the teaser as actual introduction to the article text, instead of
having to copy the introductory part of the article into the teaser.

We wrap div.realblog_teaser around the teaser in this case to allow it
to be styled to stand out.