Fix #52: entries_per_page doesn't cater to 0
Actually, we don't need a special case to suppress the pagination; this
can already be effectively done by giving a very large number. However,
we have to make sure to have a minimum of 1 (0 wouldn't make sense, and
also trigger a division by zero). And of course, we have to reflect the
behavior in the help text of the configuration option.