Close #35: Most viewed articles
We implement Realblog_mostPopular() which works rather similar to
Realblog_link(). To facilitate this list, we record every page view
including a timestamp in visitor mode without any further ado (such as
trying to exclude bots, or checking the dwelling time). We do not use
the timestamps, yet, but that might be done in the future to offer
something like most popular articles in a given period.