Refactor plugin call functions to static Realblog methods
As it is not yet possible to use static methods in plugin calls, we
automagically create the plugin call functions when Realblog is
initialized.

We mainly do this to avoid having to write and read code in the global
namespace, and maybe sometime we can get rid of the global plugin call
functions altogether.