Close #29: Category selection
We make it easier for the admin to add existing categories to an
article by offering a select element where all categories are pickable.
Picking a category will add it to the categories input without any
checking whether the category is already there. It is still the admin's
responsibility to properly set the categories input, and to delete
undesired categories manually.