Fix #51: Improve pagination
We remove the "page of pages" part of the pagination as this
information is already easily available from the actual pagination.
Furthermore we remove the • characters, as these are purely
aesthetic and as such should be left for CSS; instead we add CSS
classes to the elements.

We also refactor PaginationView to a Pagination class which uses a
proper view.