Add config option `heading_above_meta`
Some bloggers prefer to show the article's meta information below the
actual article heading; others prefer to show it above. We make that
configurable, so they neither have to edit the views (which is safe
regarding updates) nor to fiddle around with CSS (which may not be
broadly supported yet).