Stricter typing
We need to work on our error handling in the long run, but for now we
keep the current behavior by adding assertions where the code would
have thrown errors, and adding casts where that would have implicitly
happened before.  Well, mostly, since there have been a couple of minor
type bugs which we fix right away.

We also replace the dreaded ext/filter functions with manually
accessing the superglobals for now; in the long run we probably want to
introduce respective features on a Request object.

* Declare Article constructor types
* PHPStan level 9 \o/
* Drop ext/filter usage
* PHPStan level 8