Test DB and Finder
Just some rough tests to get some coverage, and to set up the machinery
to quickly add regression tests, if necessary.  We use in-memory
database which is fast and allows us to nicely isolate the tests.  To
be able to do that, we need to slightly tweak `DB::init()`.

While we're at it, we also avoid re-connecting to the DB, what should
increase the overall performance, since the DB is already shared
anyway.