Don't inject language text into controllers
Dealing with language texts is the responsibility of the view.  Period.
For "normal" cases, the controller should call the view to get the text
rendered; if necessary appropriate view methods can be added, and of
course, view templates cause almost no overhead, even if they are small.
In some other cases, the key of the language text can be passed to the
view template.

Finally, if the language text is actually a config option, it can be
injected into the plugin config.

* Extract View::json()
* Inject language specific config into plugin config
* Extract Dic::makeConf()
* Prefer calling the view to get localized texts
* Proper i18n of back-end titles
* Extract info_message template
* Extract View::message()
* Pass text *key* to view template
* Extract View::month()
* Extract Dic::makeView()
* Extract View::date()
* Extract feed_link view template