Overhaul request/response handling
We no longer pass request and response objects around in the request
handlers.  If there is the need to access either from internal methods,
we assign the objects to properties and use these.  That feels a bit
unclean, but is rather convenient, and after all, the request handlers
are method objects, so instead of initializing the objects in the
constructor, initializing them in `::__invoke()` seems to be okay.

* Don't pass request around in controllers
* Single response per request handling
* Make Response mutable