Improve default style of article overviews
First, it makes no sense to put a border-bottom on .realblog_show_story
so we move that to .realblog_entry_footer.

Then we address .realblog_read_more, where `float:right` can less
intrusively be accomplished with `text-align:right`, and where there
doesn't appear to be a need for a `padding-bottom`; after all, it's
already enclosed in a `<p>` element.